popupbug & remove test class
removed unnecessary testing class for straight (has a testcase in test folder). also fixed bug problem about several popups. The problem was the splitting of the pot into subpots if players go all in. This causes several winners in some cases, but in cases where only one user wins, it causes several popups to show up (according to the nbr of subpots).